Revert "common: init and print io stats"
This reverts commit e94e541b986080f657476303025d279f2e0bdfd3.

I/O stats will be recorded in phastaIO.cc where timers can be
placed immediately before and after the read/write/open/close
calls.  This removes the overheads and provides stats on only
those calls.  The overheads may be of interest at some point,
but for now we are just interested in bandwidth.